added random user call
