+++ app/src/main/java/com/example/randomuserapi/features/selecnumberofusers/selectnumberofusersviewmodel/SelectNumberOfUsersViewModel.kt
@@ -10,13 +10,9 @@ class SelectNumberOfUsersViewModel: SelectNumberOfUsersViewModelInterface {
 
     fun initializeViewModel(){
         randomUserUseCase = RandomUserUseCase()
-        callRandomUserUseCase()
     }
 
 
-    override fun callRandomUserUseCase(){
-        //Aquí llamamos a la recogida de datos
-        randomUserUseCase.fetchRandomUserData()
-    }
+
 
 }
