+++ app/src/main/java/com/example/randomuserapi/features/selecnumberofusers/selectnumberofusersviewmodel/SelectNumberOfUsersViewModelInterface.kt
@@ -1,5 +1,4 @@
 package com.example.randomuserapi.features.selecnumberofusers.selectnumberofusersviewmodel
 
 interface SelectNumberOfUsersViewModelInterface {
-    fun callRandomUserUseCase()
 }
